Fix a bug: Winner declared after 1st run
Previously, the game would end as soon as user inputs. This was caused
because i forgot to revert the value of userInput back to null after
the round. So, JavaScript was calculating score based on the first
input only. This was fixed by reverting the value of userInput back to
null in the checkWhoWonAndUpdateScore function.